♻️ split tool function from agent_db.py to tool_db.py
